create the component for rendering each line of the table
