+++ src/components/FilesList.vue
@@ -18,37 +18,13 @@
       </thead>
 
       <tbody>
-        <tr
+        <File
           v-for="(entry, index) in entries"
           :key="index"
+          :entry="entry"
           @mouseover="showButton(entry)"
           @mouseleave="hideButton(entry)"
-        >
-          <td><input id="checkbox" type="checkbox" /></td>
-          <td class="entry">{{ entry.name }}</td>
-          <td class="entry">{{ entry.size }}</td>
-          <td class="entry">
-            {{ entry.modification }}
-          </td>
-          <td class="download">
-            <button
-              class="download-button"
-              v-if="entry.onHover"
-              v-on:click="download(entry.filename)"
-            >
-              <DownloadIcon class="download-icon" />
-            </button>
-          </td>
-          <td class="trash">
-            <button
-              class="trash-button"
-              v-if="entry.onHover"
-              v-on:click="deleteFile(entry.filename)"
-            >
-              <TrashIcon class="trash-icon" />
-            </button>
-          </td>
-        </tr>
+        />
       </tbody>
     </div>
   </table>
@@ -56,14 +32,12 @@
 
 <script>
 import { ref } from "vue";
-import TrashIcon from "@/components/TrashIcon.vue";
-import DownloadIcon from "@/components/DownloadIcon.vue";
+import File from "@/components/File.vue";
 
 export default {
   name: "FilesList",
   components: {
-    TrashIcon,
-    DownloadIcon,
+    File,
   },
   setup() {
     // Active area for drop and upload file
@@ -105,14 +79,6 @@ export default {
     hideButton(entry) {
       entry.onHover = false;
     },
-    download(path) {
-      var msg = { type: "forward", command: "download", path: path };
-      this.$webSocketsSend(msg);
-    },
-    deleteFile(path) {
-      var msg = { type: "forward", command: "remove", path: path };
-      this.$webSocketsSend(msg);
-    },
   },
   created() {
     this.entries.forEach((element) => {
@@ -137,33 +103,12 @@ export default {
 .table {
   border-spacing: 0px;
 }
-#checkbox {
-  opacity: 0.2;
-  /* black */
 
-  border: 1px solid #0c0c0c;
-  box-sizing: border-box;
-  border-radius: 4px;
-}
 .active-dropzone {
   opacity: 0.25;
   background-color: #fff;
 }
-tbody > tr:hover td {
-  background: #f5faff;
-  height: 3rem;
-}
-.entry {
-  background: #fff;
-  height: 3rem;
-  font-family: Poppins;
-  font-style: normal;
-  font-weight: normal;
-  border-right: none;
-  border-left: none;
-  border-bottom: 1px solid rgb(231, 231, 231);
-  color: rgb(87, 87, 87);
-}
+
 .header {
   overflow: hidden;
   text-overflow: ellipsis;
@@ -198,26 +143,5 @@ tr :nth-child(3) {
 }
 tr :nth-child(4) {
   padding-right: 3rem;
-}
-
-.trash {
-  padding-left: 1rem;
-  height: 8%;
-  opacity: 0.6;
-  border-bottom: 1px solid rgb(231, 231, 231);
-  color: rgb(87, 87, 87);
-  width: 3rem;
-}
-.download {
-  padding-left: 1rem;
-  height: 8%;
-  border-bottom: 1px solid rgb(231, 231, 231);
-  color: rgb(87, 87, 87);
-  width: 3rem;
-}
-.trash-button,
-.download-button {
-  border: none;
-  background: none;
 }
 </style>
